submodule--helper: report "submodule" as our name in "-h" output
Change the rest of the usage output for "git submodule--helper" so
that it will pretend to be named "git submodule", which it already did
e.g. in the case of "status", "deinit" and "update".

This will allow us to eventually dispatch further into "git
submodule--helper" from git-submodule.sh, while emitting the correct
usage output.

This does make the usage output of "git submodule--helper" itself
"incorrect", but since nothing user-facing invokes it directly it
makes sense to continue to elide that difference entirely. A user
isn't served by us exposing this internal implementation detail.

This fixes a bug that's been there ever since "git submodule
absorbgitdirs" was added in f6f85861400 (submodule: add absorb-git-dir
function, 2016-12-12). Before this we'd emit e.g. this on invalid
usage:

	$ git submodule absorbgitdirs --blah
	error: unknown option `blah'
	usage: git submodule--helper absorbgitdirs [<options>] [<path>...]

Now we'll emit the same, except that the usage line is now referring
to the correct command run by the user:

	usage: git submodule absorbgitdirs [<options>] [<path>...]
	[.. same usage output ...]

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>